feat: turn list commands into partial classes to allow grouping
