+++ ogybot.Bot/Commands/Groups/Tome/Implementation/TomeListRemoveCommand.cs
@@ -1,34 +1,20 @@
 ﻿using Discord;
 using Discord.Interactions;
-using ogybot.Bot.Commands.Core.Validators;
-using ogybot.Bot.Handlers;
 using ogybot.Domain.Entities.UserTypes;
-using ogybot.Domain.Infrastructure.Clients;
 using ogybot.Utility.Extensions;
 
-namespace ogybot.Bot.Commands.Groups.Waitlist;
+namespace ogybot.Bot.Commands.Groups.Tome.Implementation;
 
-public sealed class WaitlistRemoveCommand : BaseWaitlistCommand
+public sealed partial class TomeListCommands
 {
-    private readonly IListCommandValidator _commandValidator;
-
-    public WaitlistRemoveCommand(
-        IBotExceptionHandler exceptionHandler,
-        IGuildClient guildClient,
-        IWaitListClient waitListClient,
-        IListCommandValidator commandValidator) : base(exceptionHandler, guildClient, waitListClient)
-    {
-        _commandValidator = commandValidator;
-    }
-
     [CommandContextType(InteractionContextType.Guild)]
-    [SlashCommand("remove", "removes a user from the wait list based on their name or index")]
-    public async Task ExecuteWaitlistRemoveCommandAsync([Summary("users-or-indexes", "The user's name or index")] string usernamesOrIndexes)
+    [SlashCommand("remove", "removes a user from the tome list based on their name or index")]
+    public async Task ExecuteRemoveCommandAsync([Summary("users-or-indexes", "The user's name or index")] string usernamesOrIndexes)
     {
-        await HandleCommandExecutionAsync(() => WaitListRemoveInstructionsAsync(usernamesOrIndexes));
+        await HandleCommandExecutionAsync(() => RemoveCommandInstructionsAsync(usernamesOrIndexes));
     }
 
-    private async Task WaitListRemoveInstructionsAsync(string usernamesOrIndexes)
+    private async Task RemoveCommandInstructionsAsync(string usernamesOrIndexes)
     {
         if (await IsInvalidContextAsync(ValidChannelId))
         {
@@ -44,7 +30,7 @@ public sealed class WaitlistRemoveCommand : BaseWaitlistCommand
             await RemovePlayerFromListAsync(usernamesOrIndexes);
         }
 
-        await FollowupAsync("Successfully removed provided player from the wait list.");
+        await FollowupAsync("Successfully removed provided player(s) from the wait list.");
     }
 
     private async Task RemoveMultiplePlayersFromListAsync(string usernamesOrIndexes)
@@ -63,7 +49,6 @@ public sealed class WaitlistRemoveCommand : BaseWaitlistCommand
 
     private async Task RemovePlayerFromListAsync(string usernameOrIndex)
     {
-
         if (short.TryParse(usernameOrIndex, out var index))
         {
             await RemoveByIndexAsync(index);
@@ -76,34 +61,34 @@ public sealed class WaitlistRemoveCommand : BaseWaitlistCommand
 
     private async Task RemoveByIndexAsync(int index)
     {
-        var list = await WaitListClient.GetListAsync(WynnGuildId);
+        var list = await TomeListClient.GetListAsync(WynnGuildId);
 
         ValidateUserBeingRemovedByIndex(index, list);
 
         // Gets the user based on the index provided. As the list count starts at 1, the index has to be subtracted by 1.
-        var waitListUser = list[index - 1];
+        var tomeListUser = list[index - 1];
 
-        await WaitListClient.RemoveUserAsync(WynnGuildId, waitListUser);
+        await TomeListClient.RemoveUserAsync(WynnGuildId, tomeListUser);
     }
 
     private async Task RemoveByNameAsync(string username)
     {
         await ValidateUserBeingRemovedByNameAsync(username);
 
-        var waitListUser = new WaitListUser(username);
+        var tomeListUser = new TomeListUser(username);
 
-        await WaitListClient.RemoveUserAsync(WynnGuildId, waitListUser);
-    }
-
-    private void ValidateUserBeingRemovedByIndex(int index, IList<WaitListUser> list)
-    {
-        _commandValidator.ValidateUserRemoval(list, index);
+        await TomeListClient.RemoveUserAsync(WynnGuildId, tomeListUser);
     }
 
     private async Task ValidateUserBeingRemovedByNameAsync(string username)
     {
-        var list = await WaitListClient.GetListAsync(WynnGuildId);
+        var list = await TomeListClient.GetListAsync(WynnGuildId);
 
         _commandValidator.ValidateUserRemoval(list, username);
+    }
+
+    private void ValidateUserBeingRemovedByIndex(int index, IList<TomeListUser> list)
+    {
+        _commandValidator.ValidateUserRemoval(list, index);
     }
 }
