+++ ogybot.Bot/Commands/Groups/Waitlist/BaseWaitlistCommand.cs
@@ -1,9 +1,11 @@
-﻿using ogybot.Bot.Commands.Core.Base;
+﻿using Discord.Interactions;
+using ogybot.Bot.Commands.Core.Base;
 using ogybot.Bot.Handlers;
 using ogybot.Domain.Infrastructure.Clients;
 
 namespace ogybot.Bot.Commands.Groups.Waitlist;
 
+[Group("waitlist", "Presents a collection of waitlist related commands.")]
 public abstract class BaseWaitlistCommand : BasePermissionRequiredCommand
 {
     protected readonly IWaitListClient WaitListClient;
