+++ ogybot.Bot/Commands/Groups/Waitlist/Implementation/WaitlistAddCommand.cs
@@ -5,32 +5,18 @@ using ogybot.Bot.Handlers;
 using ogybot.Domain.Entities.UserTypes;
 using ogybot.Domain.Infrastructure.Clients;
 
-namespace ogybot.Bot.Commands.Groups.Waitlist;
+namespace ogybot.Bot.Commands.Groups.Waitlist.Implementation;
 
-public sealed class WaitlistAddCommand : BaseWaitlistCommand
+public sealed partial class WaitlistCommands
 {
-    private readonly IListCommandValidator _commandValidator;
-
-    public WaitlistAddCommand(
-        IBotExceptionHandler exceptionHandler,
-        IGuildClient guildClient,
-        IWaitListClient waitListClient,
-        IListCommandValidator commandValidator) : base(
-        exceptionHandler,
-        guildClient,
-        waitListClient)
-    {
-        _commandValidator = commandValidator;
-    }
-
     [CommandContextType(InteractionContextType.Guild)]
     [SlashCommand("add", "Adds a user to the wait list.")]
-    public async Task ExecuteCommandAsync([Summary("user", "User to insert into the wait list")] string username)
+    public async Task ExecuteAddCommandAsync([Summary("user", "User to insert into the wait list")] string username)
     {
-        await HandleCommandExecutionAsync(() => CommandInstructionsAsync(username));
+        await HandleCommandExecutionAsync(() => AddCommandInstructionsAsync(username));
     }
 
-    private async Task CommandInstructionsAsync(string username)
+    private async Task AddCommandInstructionsAsync(string username)
     {
         if (await IsInvalidChannelAsync(ValidChannelId))
         {
