+++ ogybot.Bot/Commands/Groups/Waitlist/Implementation/WaitlistRemoveCommand.cs
@@ -6,29 +6,18 @@ using ogybot.Domain.Entities.UserTypes;
 using ogybot.Domain.Infrastructure.Clients;
 using ogybot.Utility.Extensions;
 
-namespace ogybot.Bot.Commands.Groups.Tome;
+namespace ogybot.Bot.Commands.Groups.Waitlist.Implementation;
 
-public sealed class TomeListRemoveCommand : BaseTomeCommand
+public sealed partial class WaitlistCommands
 {
-    private readonly IListCommandValidator _commandValidator;
-
-    public TomeListRemoveCommand(
-        IBotExceptionHandler exceptionHandler,
-        IGuildClient guildClient,
-        ITomeListClient tomeListClient,
-        IListCommandValidator commandValidator) : base(exceptionHandler, guildClient, tomeListClient)
-    {
-        _commandValidator = commandValidator;
-    }
-
     [CommandContextType(InteractionContextType.Guild)]
-    [SlashCommand("remove", "removes a user from the tome list based on their name or index")]
-    public async Task ExecuteCommandAsync([Summary("users-or-indexes", "The user's name or index")] string usernamesOrIndexes)
+    [SlashCommand("remove", "removes a user from the wait list based on their name or index")]
+    public async Task ExecuteRemoveCommandAsync([Summary("users-or-indexes", "The user's name or index")] string usernamesOrIndexes)
     {
-        await HandleCommandExecutionAsync(() => CommandInstructionsAsync(usernamesOrIndexes));
+        await HandleCommandExecutionAsync(() => RemoveCommandInstructionsAsync(usernamesOrIndexes));
     }
 
-    private async Task CommandInstructionsAsync(string usernamesOrIndexes)
+    private async Task RemoveCommandInstructionsAsync(string usernamesOrIndexes)
     {
         if (await IsInvalidContextAsync(ValidChannelId))
         {
@@ -44,7 +33,7 @@ public sealed class TomeListRemoveCommand : BaseTomeCommand
             await RemovePlayerFromListAsync(usernamesOrIndexes);
         }
 
-        await FollowupAsync("Successfully removed provided player(s) from the wait list.");
+        await FollowupAsync("Successfully removed provided player from the wait list.");
     }
 
     private async Task RemoveMultiplePlayersFromListAsync(string usernamesOrIndexes)
@@ -63,6 +52,7 @@ public sealed class TomeListRemoveCommand : BaseTomeCommand
 
     private async Task RemovePlayerFromListAsync(string usernameOrIndex)
     {
+
         if (short.TryParse(usernameOrIndex, out var index))
         {
             await RemoveByIndexAsync(index);
@@ -75,34 +65,34 @@ public sealed class TomeListRemoveCommand : BaseTomeCommand
 
     private async Task RemoveByIndexAsync(int index)
     {
-        var list = await TomeListClient.GetListAsync(WynnGuildId);
+        var list = await WaitListClient.GetListAsync(WynnGuildId);
 
         ValidateUserBeingRemovedByIndex(index, list);
 
         // Gets the user based on the index provided. As the list count starts at 1, the index has to be subtracted by 1.
-        var tomeListUser = list[index - 1];
+        var waitListUser = list[index - 1];
 
-        await TomeListClient.RemoveUserAsync(WynnGuildId, tomeListUser);
+        await WaitListClient.RemoveUserAsync(WynnGuildId, waitListUser);
     }
 
     private async Task RemoveByNameAsync(string username)
     {
         await ValidateUserBeingRemovedByNameAsync(username);
 
-        var tomeListUser = new TomeListUser(username);
+        var waitListUser = new WaitListUser(username);
 
-        await TomeListClient.RemoveUserAsync(WynnGuildId, tomeListUser);
+        await WaitListClient.RemoveUserAsync(WynnGuildId, waitListUser);
+    }
+
+    private void ValidateUserBeingRemovedByIndex(int index, IList<WaitListUser> list)
+    {
+        _commandValidator.ValidateUserRemoval(list, index);
     }
 
     private async Task ValidateUserBeingRemovedByNameAsync(string username)
     {
-        var list = await TomeListClient.GetListAsync(WynnGuildId);
+        var list = await WaitListClient.GetListAsync(WynnGuildId);
 
         _commandValidator.ValidateUserRemoval(list, username);
-    }
-
-    private void ValidateUserBeingRemovedByIndex(int index, IList<TomeListUser> list)
-    {
-        _commandValidator.ValidateUserRemoval(list, index);
     }
 }
